Extracted test_suite_subcases
